pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opticlient"
version = "0.1.5"
description = "Python client for SaaS optimization API"
readme = { file = "DESCRIPTION.md", content-type = "text/markdown" }
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
  { name = "Milan Adhikari", email = "reach.out.to.milan@gmail.com" },
]

dependencies = [
  "httpx",
  "python-dotenv>=1.2.1"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "build",
  "twine",
]

[project.urls]
Homepage = "https://cad-eta.vercel.app"
Source = "https://github.com/Milan-Adhikari/opticlient"
